+++ delivery-agent-service/src/main/java/com/eighteenthstreet/deliveryagentservice/domain/event/DeliveryFailedEvent.java
@@ -1,0 +1,8 @@
+package com.eighteenthstreet.deliveryagentservice.domain.event;
+
+import java.util.UUID;
+
+import exception.ErrorCode;
+
+public record DeliveryFailedEvent(UUID deliveryId, ErrorCode errorCode) {
+}
